criacao dos outros endpoints

+++ src/main/java/com/rinha/dto/PaymentSummary.java
@@ -2,5 +2,5 @@ package com.rinha.dto;
 
 import java.math.BigDecimal;
 
-public record PaymentSummary(Long totalRequests, BigDecimal totalAmount) {
+public record PaymentSummary(Integer totalRequests, BigDecimal totalAmount) {
 }
